refactor: introduce an abstract saml plugin to reduce code duplication
