+++ wrapper/src/main/java/software/amazon/jdbc/Driver.java
@@ -42,8 +42,7 @@ import software.amazon.jdbc.plugin.DataCacheConnectionPlugin;
 import software.amazon.jdbc.plugin.OpenedConnectionTracker;
 import software.amazon.jdbc.plugin.customendpoint.CustomEndpointMonitorImpl;
 import software.amazon.jdbc.plugin.efm.HostMonitorThreadContainer;
-import software.amazon.jdbc.plugin.federatedauth.FederatedAuthCacheHolder;
-import software.amazon.jdbc.plugin.federatedauth.OktaAuthCacheHolder;
+import software.amazon.jdbc.plugin.federatedauth.AuthCacheHolder;
 import software.amazon.jdbc.plugin.iam.IamAuthCacheHolder;
 import software.amazon.jdbc.plugin.limitless.LimitlessRouterServiceImpl;
 import software.amazon.jdbc.profile.ConfigurationProfile;
@@ -448,8 +447,7 @@ public class Driver implements java.sql.Driver {
     OpenedConnectionTracker.clearCache();
     AwsSecretsManagerCacheHolder.clearCache();
     DataCacheConnectionPlugin.clearCache();
-    FederatedAuthCacheHolder.clearCache();
-    OktaAuthCacheHolder.clearCache();
+    AuthCacheHolder.clearCache();
     IamAuthCacheHolder.clearCache();
     LimitlessRouterServiceImpl.clearCache();
     RoundRobinHostSelector.clearCache();
