+++ wrapper/src/main/java/software/amazon/jdbc/plugin/federatedauth/AuthCacheHolder.java
@@ -19,11 +19,7 @@ package software.amazon.jdbc.plugin.federatedauth;
 import java.util.concurrent.ConcurrentHashMap;
 import software.amazon.jdbc.plugin.TokenInfo;
 
-/* The main plugin code FederatedAuthPlugin depends on AWS SDK. In order to avoid unnecessary dependencies,
- the plugin cache has been extracted into this FederatedAuthCacheHolder class. This cache holder class doesn't depend
- on AWS SDK and can be safely cleared if needed.
- */
-public class FederatedAuthCacheHolder {
+public class AuthCacheHolder {
   static final ConcurrentHashMap<String, TokenInfo> tokenCache = new ConcurrentHashMap<>();
 
   public static void clearCache() {
