+++ wrapper/src/main/java/software/amazon/jdbc/plugin/federatedauth/FederatedAuthPlugin.java
@@ -16,41 +16,23 @@
 
 package software.amazon.jdbc.plugin.federatedauth;
 
-import java.sql.Connection;
-import java.sql.SQLException;
-import java.time.Instant;
-import java.time.temporal.ChronoUnit;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.Properties;
 import java.util.Set;
 import java.util.logging.Logger;
 import java.util.regex.Pattern;
-import org.checkerframework.checker.nullness.qual.NonNull;
-import software.amazon.awssdk.auth.credentials.AwsCredentialsProvider;
-import software.amazon.awssdk.regions.Region;
 import software.amazon.jdbc.AwsWrapperProperty;
-import software.amazon.jdbc.HostSpec;
-import software.amazon.jdbc.JdbcCallable;
 import software.amazon.jdbc.PluginService;
 import software.amazon.jdbc.PropertyDefinition;
-import software.amazon.jdbc.plugin.AbstractConnectionPlugin;
-import software.amazon.jdbc.plugin.TokenInfo;
 import software.amazon.jdbc.plugin.iam.IamTokenUtility;
-import software.amazon.jdbc.util.GDBRegionUtils;
 import software.amazon.jdbc.util.IamAuthUtils;
 import software.amazon.jdbc.util.Messages;
-import software.amazon.jdbc.util.RdsUrlType;
 import software.amazon.jdbc.util.RdsUtils;
-import software.amazon.jdbc.util.RegionUtils;
-import software.amazon.jdbc.util.StringUtils;
 import software.amazon.jdbc.util.telemetry.TelemetryCounter;
-import software.amazon.jdbc.util.telemetry.TelemetryFactory;
 import software.amazon.jdbc.util.telemetry.TelemetryGauge;
 
-public class FederatedAuthPlugin extends AbstractConnectionPlugin {
-
-  private final CredentialsProviderFactory credentialsProviderFactory;
+public class FederatedAuthPlugin extends BaseSamlAuthPlugin {
   private static final int DEFAULT_TOKEN_EXPIRATION_SEC = 15 * 60 - 30;
   private static final int DEFAULT_HTTP_TIMEOUT_MILLIS = 60000;
   public static final AwsWrapperProperty IDP_ENDPOINT = new AwsWrapperProperty("idpEndpoint", null,
@@ -90,13 +72,6 @@ public class FederatedAuthPlugin extends AbstractConnectionPlugin {
       new AwsWrapperProperty("dbUser", null, "The database user used to access the database");
   protected static final Pattern SAML_RESPONSE_PATTERN = Pattern.compile("SAMLResponse\\W+value=\"(?<saml>[^\"]+)\"");
   protected static final String SAML_RESPONSE_PATTERN_GROUP = "saml";
-  private static final Logger LOGGER = Logger.getLogger(FederatedAuthPlugin.class.getName());
-
-  protected final PluginService pluginService;
-
-  protected final RdsUtils rdsUtils;
-  protected final SamlUtils samlUtils;
-  protected RegionUtils regionUtils;
 
   private static final Set<String> subscribedMethods =
       Collections.unmodifiableSet(new HashSet<String>() {
@@ -110,11 +85,8 @@ public class FederatedAuthPlugin extends AbstractConnectionPlugin {
     PropertyDefinition.registerPluginProperties(FederatedAuthPlugin.class);
   }
 
-  private final TelemetryFactory telemetryFactory;
   private final TelemetryGauge cacheSizeGauge;
   private final TelemetryCounter fetchTokenCounter;
-  private final IamTokenUtility iamTokenUtility;
-
 
   @Override
   public Set<String> getSubscribedMethods() {
@@ -131,6 +103,7 @@ public class FederatedAuthPlugin extends AbstractConnectionPlugin {
       final CredentialsProviderFactory credentialsProviderFactory,
       final RdsUtils rdsUtils,
       final IamTokenUtility tokenUtils) {
+    super(pluginService, credentialsProviderFactory, rdsUtils, tokenUtils);
     try {
       Class.forName("software.amazon.awssdk.services.sts.model.AssumeRoleWithSamlRequest");
     } catch (final ClassNotFoundException e) {
@@ -141,150 +114,44 @@ public class FederatedAuthPlugin extends AbstractConnectionPlugin {
       }
     }
 
-    this.pluginService = pluginService;
-    this.credentialsProviderFactory = credentialsProviderFactory;
-    this.telemetryFactory = pluginService.getTelemetryFactory();
     this.cacheSizeGauge = this.telemetryFactory.createGauge("federatedAuth.tokenCache.size",
-        () -> (long) FederatedAuthCacheHolder.tokenCache.size());
+        () -> (long) AuthCacheHolder.tokenCache.size());
     this.fetchTokenCounter = this.telemetryFactory.createCounter("federatedAuth.fetchToken.count");
-    this.rdsUtils = rdsUtils;
-    this.samlUtils = new SamlUtils(this.rdsUtils);
-    this.iamTokenUtility = tokenUtils;
+  }
+
+
+  @Override
+  String getDbUserProperty(Properties props) {
+    return DB_USER.getString(props);
   }
 
   @Override
-  public Connection connect(
-      final String driverProtocol,
-      final HostSpec hostSpec,
-      final Properties props,
-      final boolean isInitialConnection,
-      final JdbcCallable<Connection, SQLException> connectFunc)
-      throws SQLException {
-    return connectInternal(hostSpec, props, connectFunc);
+  String getIamHostProperty(Properties props) {
+    return IAM_HOST.getString(props);
   }
 
   @Override
-  public Connection forceConnect(
-      final @NonNull String driverProtocol,
-      final @NonNull HostSpec hostSpec,
-      final @NonNull Properties props,
-      final boolean isInitialConnection,
-      final @NonNull JdbcCallable<Connection, SQLException> forceConnectFunc)
-      throws SQLException {
-    return connectInternal(hostSpec, props, forceConnectFunc);
-  }
-
-  private Connection connectInternal(final HostSpec hostSpec, final Properties props,
-      final JdbcCallable<Connection, SQLException> connectFunc) throws SQLException {
-
-    this.samlUtils.checkIdpCredentialsWithFallback(IDP_USERNAME, IDP_PASSWORD, props);
-
-    final HostSpec host = IamAuthUtils.getIamHost(IAM_HOST.getString(props), hostSpec);
-
-    final int port = IamAuthUtils.getIamPort(
-        IAM_DEFAULT_PORT.getInteger(props),
-        hostSpec,
-        this.pluginService.getDialect().getDefaultPort());
-
-    final RdsUrlType type = rdsUtils.identifyRdsType(host.getHost());
-
-    AwsCredentialsProvider credentialsProvider = null;
-    if (RdsUrlType.RDS_GLOBAL_WRITER_CLUSTER == type) {
-      // Assume a role early so we can use this credential to fetch region information.
-      credentialsProvider = this.credentialsProviderFactory.getAwsCredentialsProvider(hostSpec.getHost(), null, props);
-      this.regionUtils = new GDBRegionUtils(credentialsProvider);
-    } else {
-      this.regionUtils = new RegionUtils();
-    }
-
-    final Region region = this.regionUtils.getRegion(host, props, IAM_REGION.name);
-    if (region == null) {
-      throw new SQLException(
-          Messages.get("FederatedAuthPlugin.unableToDetermineRegion", new Object[]{ IAM_REGION.name }));
-    }
-
-    final String cacheKey = IamAuthUtils.getCacheKey(
-        DB_USER.getString(props),
-        host.getHost(),
-        port,
-        region);
-
-    final TokenInfo tokenInfo = FederatedAuthCacheHolder.tokenCache.get(cacheKey);
-
-    final boolean isCachedToken = tokenInfo != null && !tokenInfo.isExpired();
-
-    if (isCachedToken) {
-      LOGGER.finest(
-          () -> Messages.get(
-              "AuthenticationToken.useCachedToken",
-              new Object[] {tokenInfo.getToken()}));
-      PropertyDefinition.PASSWORD.set(props, tokenInfo.getToken());
-    } else {
-      updateAuthenticationToken(hostSpec, props, region, cacheKey, host.getHost(), credentialsProvider);
-    }
-
-    PropertyDefinition.USER.set(props, DB_USER.getString(props));
-
+  int getIamPortProperty(Properties props) {
     try {
-      return connectFunc.call();
-    } catch (final SQLException exception) {
-      if (!isCachedToken
-          || !this.pluginService.isLoginException(exception, this.pluginService.getTargetDriverDialect())) {
-        throw exception;
-      }
-      updateAuthenticationToken(hostSpec, props, region, cacheKey, host.getHost(), credentialsProvider);
-      return connectFunc.call();
-    } catch (final Exception exception) {
-      LOGGER.warning(
-          () -> Messages.get(
-              "SamlAuthPlugin.unhandledException",
-              new Object[] {exception}));
-      throw new SQLException(exception);
+      return IAM_DEFAULT_PORT.getInteger(props);
+    } catch (NumberFormatException e) {
+      // Return 0 if IAM_DEFAULT_PORT returns null or an empty string.
+      return 0;
     }
   }
 
-  private void updateAuthenticationToken(
-      final HostSpec hostSpec,
-      final Properties props,
-      final Region region,
-      final String cacheKey,
-      final String host,
-      AwsCredentialsProvider credentialsProvider)
-      throws SQLException {
-    final int tokenExpirationSec = IAM_TOKEN_EXPIRATION.getInteger(props);
-    final Instant tokenExpiry = Instant.now().plus(tokenExpirationSec, ChronoUnit.SECONDS);
-    final int port = IamAuthUtils.getIamPort(
-        StringUtils.isNullOrEmpty(IAM_DEFAULT_PORT.getString(props)) ? 0 : IAM_DEFAULT_PORT.getInteger(props),
-        hostSpec,
-        this.pluginService.getDialect().getDefaultPort());
-
-    if (credentialsProvider == null) {
-      credentialsProvider =
-          this.credentialsProviderFactory.getAwsCredentialsProvider(hostSpec.getHost(), region, props);
+  @Override
+  int getIamTokenExpiration(Properties props) {
+    try {
+      return IAM_TOKEN_EXPIRATION.getInteger(props);
+    } catch (NumberFormatException e) {
+      // Return 0 if IAM_DEFAULT_PORT returns null or an empty string.
+      return DEFAULT_TOKEN_EXPIRATION_SEC;
     }
-
-    if (this.fetchTokenCounter != null) {
-      this.fetchTokenCounter.inc();
-    }
-    final String token = IamAuthUtils.generateAuthenticationToken(
-        this.iamTokenUtility,
-        this.pluginService,
-        DB_USER.getString(props),
-        host,
-        port,
-        region,
-        credentialsProvider);
-    LOGGER.finest(
-        () -> Messages.get(
-            "AuthenticationToken.generatedNewToken",
-            new Object[] {token}));
-    PropertyDefinition.PASSWORD.set(props, token);
-    FederatedAuthCacheHolder.tokenCache.put(
-        cacheKey,
-        new TokenInfo(token, tokenExpiry));
   }
 
-  public static void clearCache() {
-    FederatedAuthCacheHolder.clearCache();
+  @Override
+  void checkIdpCredentialsWithFallback(Properties props) {
+    this.samlUtils.checkIdpCredentialsWithFallback(IDP_USERNAME, IDP_PASSWORD, props);
   }
 }
